- further implementation of validationMethods

+++ api/parseAuctionData.php
@@ -3,7 +3,7 @@
 require_once '../dependencies/headers.php';
 require_once '../dependencies/class.AuctionCraftSniper.php';
 
-$decodedPOST = json_decode(trim(file_get_contents("php://input")), true);
+$decodedPOST = json_decode (trim (file_get_contents ("php://input")), true);
 
 const BYTE_LIMIT = 365;
 const CHUNK_SIZE = 2568505;
@@ -12,39 +12,39 @@ if (!isset($decodedPOST['step']) && empty($decodedPOST['recipeIDs'])) {
 
     $AuctionCraftSniper = new AuctionCraftSniper();
 
-    $expansionLevel = (int)$decodedPOST['expansionLevel'];
+    $expansionLevel = $AuctionCraftSniper->isValidExpansionLevel ((int) $decodedPOST['expansionLevel']);
 
-    if ($expansionLevel === 0) {
+    if ($expansionLevel && $expansionLevel === 0) {
         $expansionLevel = 8;
     }
 
-    $recipeIDs = $AuctionCraftSniper->getRecipeIDs($expansionLevel);
+    $recipeIDs = $AuctionCraftSniper->getRecipeIDs ($expansionLevel);
     $step      = 0;
 
 } else {
 
-    $recipeIDs      = (array)$decodedPOST['recipeIDs'];
-    $step           = (int)$decodedPOST['step'];
-    $expansionLevel = (int)$decodedPOST['expansionLevel'];
+    $recipeIDs      = (array) $decodedPOST['recipeIDs'];
+    $step           = (int) $decodedPOST['step'];
+    $expansionLevel = (int) $decodedPOST['expansionLevel'];
 
     if ($step < 1) {
-        echo json_encode(['err' => 'invalid step specified']);
+        echo json_encode (['err' => 'invalid step specified']);
         die;
     }
 }
 
-$fileName = (int)$decodedPOST['house'] . '.json';
+$fileName = (int) $decodedPOST['house'] . '.json';
 
 if ($fileName === '0.json') {
-    echo json_encode(['err' => 'invalid house specified']);
+    echo json_encode (['err' => 'invalid house specified']);
     die;
 }
 
-if (file_exists($fileName) && $stream = fopen($fileName, 'r')) {
+if (file_exists ($fileName) && $stream = fopen ($fileName, 'r')) {
 
     if (!isset($decodedPOST['step'])) {
-        $first200Bytes = stream_get_contents($stream, 200, 0);
-        $auctionsStart = strpos($first200Bytes, '"auctions": [') + 16;
+        $first200Bytes = stream_get_contents ($stream, 200, 0);
+        $auctionsStart = strpos ($first200Bytes, '"auctions": [') + 16;
 
         $thisChunksEnd = CHUNK_SIZE;
     } else {
@@ -52,7 +52,7 @@ if (file_exists($fileName) && $stream = fopen($fileName, 'r')) {
         $auctionsStart = $thisChunksEnd - CHUNK_SIZE;
     }
 
-    $fileSize = filesize($fileName);
+    $fileSize = filesize ($fileName);
 
     $auctionEndString = ',{"auc"';
 
@@ -69,7 +69,7 @@ if (file_exists($fileName) && $stream = fopen($fileName, 'r')) {
         $data = $leftovers;
 
         // remove whitespace & linebreaks
-        $data .= str_replace("	", '', str_replace("\r\n", '', stream_get_contents($stream, BYTE_LIMIT, $bytes)));
+        $data .= str_replace ("	", '', str_replace ("\r\n", '', stream_get_contents ($stream, BYTE_LIMIT, $bytes)));
 
         // switch $auctionEndString since the very last auction obviously has no object following
         if ($bytes + BYTE_LIMIT >= $fileSize) {
@@ -77,17 +77,17 @@ if (file_exists($fileName) && $stream = fopen($fileName, 'r')) {
         }
 
         // find end of this auction
-        $auctionEnd = strpos($data, $auctionEndString);
+        $auctionEnd = strpos ($data, $auctionEndString);
 
         // define new leftovers for next iteration, +1 because of leading , at start of new auction obj
-        $leftovers = substr($data, $auctionEnd + 1);
+        $leftovers = substr ($data, $auctionEnd + 1);
 
-        $data = json_decode(substr($data, 0, $auctionEnd), true);
+        $data = json_decode (substr ($data, 0, $auctionEnd), true);
 
-        if (in_array($data['item'], array_keys($recipeIDs))) {
-            $thisPPU = round($data['buyout'] / $data['quantity']);
+        if (in_array ($data['item'], array_keys ($recipeIDs))) {
+            $thisPPU = round ($data['buyout'] / $data['quantity']);
 
-            $previousPPU = (int)$recipeIDs[$data['item']];
+            $previousPPU = (int) $recipeIDs[$data['item']];
 
             if ($previousPPU === 0 || $thisPPU < $previousPPU && $thisPPU !== 0) {
                 $recipeIDs[$data['item']] = $thisPPU;
@@ -95,24 +95,27 @@ if (file_exists($fileName) && $stream = fopen($fileName, 'r')) {
         }
     }
 
-    fclose($stream);
+    fclose ($stream);
 
     $response = [
         'recipeIDs'      => $recipeIDs,
         'expansionLevel' => $expansionLevel,
         'step'           => $step + 1,
-        'reqSteps'       => ceil($fileSize / CHUNK_SIZE),
-        'percentDone'    => round(($thisChunksEnd / $fileSize) * 100, 2),
+        'reqSteps'       => ceil ($fileSize / CHUNK_SIZE),
+        'percentDone'    => round (($thisChunksEnd / $fileSize) * 100, 2),
     ];
 
-    if ((int)$response['step'] === (int)$response['reqSteps']) {
+    if ((int) $response['step'] === (int) $response['reqSteps']) {
         $AuctionCraftSniper = new AuctionCraftSniper();
-        $AuctionCraftSniper->updateHouse($decodedPOST['house'], $recipeIDs, $expansionLevel);
 
-        $response['callback'] = 'getProfessionTables';
-        unlink($fileName);
+        if ($AuctionCraftSniper->isValidExpansionLevel ($expansionLevel) && $AuctionCraftSniper->isValidHouse ($decodedPOST['house'])) {
+            $AuctionCraftSniper->updateHouse ($decodedPOST['house'], $recipeIDs, $expansionLevel);
+
+            $response['callback'] = 'getProfessionTables';
+            unlink ($fileName);
+        }
     }
 
-    echo json_encode($response);
+    echo json_encode ($response);
 }
